Splitt to komponentar til kvar si fil
