+++ src/components/modal/tildel-veileder/tildel-veileder.tsx
@@ -1,25 +1,19 @@
-import * as React from 'react';
-import {useState} from 'react';
+import React, {useState} from 'react';
 import {connect, useDispatch, useSelector} from 'react-redux';
+import {BodyShort, Button, Heading, Modal} from '@navikt/ds-react';
 import {tildelVeileder} from '../../../ducks/portefolje';
-import {BrukerModell, VeilederModell} from '../../../model-interfaces';
+import {BrukerModell} from '../../../model-interfaces';
 import {AppState} from '../../../reducer';
-import '../../toolbar/toolbar.css';
 import SokFilter from '../../sok-veiledere/sok-filter';
-import classNames from 'classnames';
 import {nameToStateSliceMap} from '../../../ducks/utils';
 import {useSelectGjeldendeVeileder} from '../../../hooks/portefolje/use-select-gjeldende-veileder';
-import {BodyShort, Button, Heading, Modal, Radio, RadioGroup} from '@navikt/ds-react';
 import {useIdentSelector} from '../../../hooks/redux/use-innlogget-ident';
 import {Fnr, FnrList} from '../../fnr-list';
 import {useEnhetSelector} from '../../../hooks/redux/use-enhet-selector';
 import {trackAmplitude} from '../../../amplitude/amplitude';
 import {OversiktType} from '../../../ducks/ui/listevisning';
-
-interface TildelVeilederProps {
-    oversiktType?: OversiktType;
-    closeInput: () => void;
-}
+import {TildelVeilederRenderer} from './tildel-veileder-renderer';
+import '../../toolbar/toolbar.css';
 
 const fjernduplikatOgMapTilFnrArray = (brukereSomTildeles: BrukerModell[]) =>
     brukereSomTildeles.reduce((arrayUtenDuplikater: Fnr[], bruker: BrukerModell) => {
@@ -31,6 +25,11 @@ const fjernduplikatOgMapTilFnrArray = (brukereSomTildeles: BrukerModell[]) =>
         });
         return arrayUtenDuplikater;
     }, []);
+
+interface TildelVeilederProps {
+    oversiktType?: OversiktType;
+    closeInput: () => void;
+}
 
 function TildelVeileder({oversiktType, closeInput}: TildelVeilederProps) {
     const [ident, setIdent] = useState<string | null>(null);
@@ -226,47 +225,6 @@ function TildelVeileder({oversiktType, closeInput}: TildelVeilederProps) {
                 )}
             </SokFilter>
         </>
-    );
-}
-
-interface TildelVeilederRendererProps {
-    data: VeilederModell[];
-    ident: string | null;
-    onChange: (ident: string) => void;
-    btnOnClick: () => void;
-    oversiktType: OversiktType | undefined;
-}
-
-function TildelVeilederRenderer({data, ident, onChange, btnOnClick}: TildelVeilederRendererProps) {
-    return (
-        <form className="skjema radio-filterform" data-testid="tildel-veileder_dropdown">
-            <RadioGroup hideLegend legend="" className="radio-filterform__valg" onChange={onChange}>
-                {data.map((veileder, index) => (
-                    <Radio
-                        data-testid={`tildel-veileder_valg_${index}`}
-                        type={'button'}
-                        key={veileder.ident}
-                        name="veileder"
-                        size="small"
-                        value={veileder.ident}
-                        className={'navds-radio'}
-                    >{`${veileder.etternavn}, ${veileder.fornavn}`}</Radio>
-                ))}
-            </RadioGroup>
-            <div className="filterform__under-valg">
-                <Button
-                    size="small"
-                    onClick={btnOnClick}
-                    className={classNames('knapp', 'knapp--mini', {
-                        'knapp--hoved': ident
-                    })}
-                    type={'button'}
-                    data-testid={ident ? 'tildel-veileder_velg-knapp' : 'tildel-veileder_lukk-knapp'}
-                >
-                    {ident ? 'Velg' : 'Lukk'}
-                </Button>
-            </div>
-        </form>
     );
 }
